requirements.txt
fastapi
uvicorn
httpx
python-jose[cryptography]
python-dotenv
requests
